+++ processor/src/fast/memory.rs
@@ -3,7 +3,7 @@ use alloc::{collections::BTreeMap, vec::Vec};
 use miden_air::RowIndex;
 use vm_core::{EMPTY_WORD, Felt, WORD_SIZE, Word, ZERO};
 
-use crate::{ContextId, MemoryAddress, MemoryError};
+use crate::{ContextId, ErrorContext, MemoryAddress, MemoryError};
 
 /// The memory for the processor.
 ///
@@ -24,8 +24,14 @@ impl Memory {
     ///
     /// # Errors
     /// - Returns an error if the provided address is out-of-bounds.
-    pub fn read_element(&mut self, ctx: ContextId, addr: Felt) -> Result<Felt, MemoryError> {
-        let element = self.read_element_impl(ctx, clean_addr(addr)?).unwrap_or(ZERO);
+    #[inline(always)]
+    pub fn read_element(
+        &mut self,
+        ctx: ContextId,
+        addr: Felt,
+        err_ctx: &impl ErrorContext,
+    ) -> Result<Felt, MemoryError> {
+        let element = self.read_element_impl(ctx, clean_addr(addr, err_ctx)?).unwrap_or(ZERO);
         Ok(element)
     }
 
@@ -33,14 +39,16 @@ impl Memory {
     ///
     /// # Errors
     /// - Returns an error if the provided address is out-of-bounds or not word-aligned.
+    #[inline(always)]
     pub fn read_word(
         &self,
         ctx: ContextId,
         addr: Felt,
         clk: RowIndex,
+        err_ctx: &impl ErrorContext,
     ) -> Result<Word, MemoryError> {
-        let addr = clean_addr(addr)?;
-        let word = self.read_word_impl(ctx, addr, Some(clk))?.unwrap_or(EMPTY_WORD);
+        let addr = clean_addr(addr, err_ctx)?;
+        let word = self.read_word_impl(ctx, addr, Some(clk), err_ctx)?.unwrap_or(EMPTY_WORD);
 
         Ok(word)
     }
@@ -49,13 +57,15 @@ impl Memory {
     ///
     /// # Errors
     /// - Returns an error if the provided address is out-of-bounds.
+    #[inline(always)]
     pub fn write_element(
         &mut self,
         ctx: ContextId,
         addr: Felt,
         element: Felt,
+        err_ctx: &impl ErrorContext,
     ) -> Result<(), MemoryError> {
-        let (word_addr, idx) = split_addr(clean_addr(addr)?);
+        let (word_addr, idx) = split_addr(clean_addr(addr, err_ctx)?);
 
         self.memory
             .entry((ctx, word_addr))
@@ -78,14 +88,16 @@ impl Memory {
     ///
     /// # Errors
     /// - Returns an error if the provided address is out-of-bounds or not word-aligned.
+    #[inline(always)]
     pub fn write_word(
         &mut self,
         ctx: ContextId,
         addr: Felt,
         clk: RowIndex,
         word: Word,
+        err_ctx: &impl ErrorContext,
     ) -> Result<(), MemoryError> {
-        let addr = enforce_word_aligned_addr(ctx, clean_addr(addr)?, Some(clk))?;
+        let addr = enforce_word_aligned_addr(ctx, clean_addr(addr, err_ctx)?, Some(clk), err_ctx)?;
         self.memory.insert((ctx, addr), word);
 
         Ok(())
@@ -130,13 +142,15 @@ impl Memory {
     /// # Returns
     /// - The word starting at the provided address, if it was written previously.
     /// - `None` if the memory was not written previously.
+    #[inline(always)]
     pub(crate) fn read_word_impl(
         &self,
         ctx: ContextId,
         addr: u32,
         clk: Option<RowIndex>,
+        err_ctx: &impl ErrorContext,
     ) -> Result<Option<Word>, MemoryError> {
-        let addr = enforce_word_aligned_addr(ctx, addr, clk)?;
+        let addr = enforce_word_aligned_addr(ctx, addr, clk, err_ctx)?;
         let word = self.memory.get(&(ctx, addr)).copied();
 
         Ok(word)
@@ -150,9 +164,10 @@ impl Memory {
 ///
 /// # Errors
 /// - Returns an error if the provided address is out-of-bounds.
-fn clean_addr(addr: Felt) -> Result<u32, MemoryError> {
+#[inline(always)]
+fn clean_addr(addr: Felt, err_ctx: &impl ErrorContext) -> Result<u32, MemoryError> {
     let addr = addr.as_int();
-    addr.try_into().map_err(|_| MemoryError::address_out_of_bounds(addr, &()))
+    addr.try_into().map_err(|_| MemoryError::address_out_of_bounds(addr, err_ctx))
 }
 
 /// Splits the provided address into the word address and the index within the word.
@@ -171,16 +186,21 @@ fn split_addr(addr: u32) -> (u32, u32) {
 /// # Errors
 /// - Returns an error if the provided address is not word-aligned.
 /// - Returns an error if the provided address is out-of-bounds.
+#[inline(always)]
 fn enforce_word_aligned_addr(
     ctx: ContextId,
     addr: u32,
     clk: Option<RowIndex>,
+    err_ctx: &impl ErrorContext,
 ) -> Result<u32, MemoryError> {
     if addr % WORD_SIZE as u32 != 0 {
         return match clk {
-            Some(clk) => {
-                Err(MemoryError::unaligned_word_access(addr, ctx, Felt::from(clk.as_u32()), &()))
-            },
+            Some(clk) => Err(MemoryError::unaligned_word_access(
+                addr,
+                ctx,
+                Felt::from(clk.as_u32()),
+                err_ctx,
+            )),
             None => Err(MemoryError::UnalignedWordAccessNoClk { addr, ctx }),
         };
     }