+++ processor/src/fast/mod.rs
@@ -1,4 +1,4 @@
-use alloc::vec::Vec;
+use alloc::{sync::Arc, vec::Vec};
 use core::cmp::min;
 
 use memory::Memory;
@@ -6,6 +6,7 @@ use miden_air::RowIndex;
 use vm_core::{
     Decorator, DecoratorIterator, EMPTY_WORD, Felt, Kernel, ONE, Operation, Program, StackOutputs,
     WORD_SIZE, Word, ZERO,
+    debuginfo::{DefaultSourceManager, SourceManager},
     mast::{
         BasicBlockNode, CallNode, DynNode, ExternalNode, JoinNode, LoopNode, MastForest, MastNode,
         MastNodeId, OP_GROUP_SIZE, OpBatch, SplitNode,
@@ -15,8 +16,8 @@ use vm_core::{
 };
 
 use crate::{
-    ContextId, ExecutionError, FMP_MIN, Host, ProcessState, SYSCALL_FMP_MIN, chiplets::Ace,
-    utils::resolve_external_node,
+    ContextId, ErrorContext, ExecutionError, FMP_MIN, Host, ProcessState, SYSCALL_FMP_MIN,
+    add_error_ctx_to_external_error, chiplets::Ace, err_ctx, utils::resolve_external_node,
 };
 
 mod memory;
@@ -141,6 +142,9 @@ pub struct FastProcessor {
 
     /// Whether to enable debug statements and tracing.
     in_debug_mode: bool,
+
+    /// The source manager (providing information about the location of each instruction).
+    source_manager: Arc<dyn SourceManager>,
 }
 
 impl FastProcessor {
@@ -186,7 +190,7 @@ impl FastProcessor {
         let stack_bot_idx = stack_top_idx - MIN_STACK_DEPTH;
 
         let bounds_check_counter = stack_bot_idx;
-
+        let source_manager = Arc::new(DefaultSourceManager::default());
         FastProcessor {
             stack,
             stack_top_idx,
@@ -201,7 +205,14 @@ impl FastProcessor {
             call_stack: Vec::new(),
             ace: Ace::default(),
             in_debug_mode,
+            source_manager,
         }
+    }
+
+    /// Set the internal source manager to an externally initialized one.
+    pub fn with_source_manager(mut self, source_manager: Arc<dyn SourceManager>) -> Self {
+        self.source_manager = source_manager;
+        self
     }
 
     // ACCESSORS
@@ -367,9 +378,19 @@ impl FastProcessor {
                 self.execute_loop_node(loop_node, program, kernel, host)?
             },
             MastNode::Call(call_node) => {
-                self.execute_call_node(call_node, program, kernel, host)?
+                let err_ctx = err_ctx!(program, call_node, self.source_manager.clone());
+                add_error_ctx_to_external_error(
+                    self.execute_call_node(call_node, program, kernel, host),
+                    err_ctx,
+                )?
             },
-            MastNode::Dyn(dyn_node) => self.execute_dyn_node(dyn_node, program, kernel, host)?,
+            MastNode::Dyn(dyn_node) => {
+                let err_ctx = err_ctx!(program, dyn_node, self.source_manager.clone());
+                add_error_ctx_to_external_error(
+                    self.execute_dyn_node(dyn_node, program, kernel, host),
+                    err_ctx,
+                )?
+            },
             MastNode::External(external_node) => {
                 self.execute_external_node(external_node, kernel, host)?
             },
@@ -422,7 +443,8 @@ impl FastProcessor {
         } else if condition == ZERO {
             self.execute_mast_node(split_node.on_false(), program, kernel, host)
         } else {
-            Err(ExecutionError::not_binary_value_if(condition, &()))
+            let err_ctx = err_ctx!(program, split_node, self.source_manager.clone());
+            Err(ExecutionError::not_binary_value_if(condition, &err_ctx))
         };
 
         // Corresponds to the row inserted for the END operation added to the trace.
@@ -470,7 +492,8 @@ impl FastProcessor {
         if condition == ZERO {
             Ok(())
         } else {
-            Err(ExecutionError::not_binary_value_loop(condition, &()))
+            let err_ctx = err_ctx!(program, loop_node, self.source_manager.clone());
+            Err(ExecutionError::not_binary_value_loop(condition, &err_ctx))
         }
     }
 
@@ -481,6 +504,8 @@ impl FastProcessor {
         kernel: &Kernel,
         host: &mut impl Host,
     ) -> Result<(), ExecutionError> {
+        let err_ctx = err_ctx!(program, call_node, self.source_manager.clone());
+
         // Corresponds to the row inserted for the CALL or SYSCALL operation added to the trace.
         self.clk += 1_u32;
 
@@ -500,7 +525,7 @@ impl FastProcessor {
         if call_node.is_syscall() {
             // check if the callee is in the kernel
             if !kernel.contains_proc(callee_hash) {
-                return Err(ExecutionError::syscall_target_not_in_kernel(callee_hash, &()));
+                return Err(ExecutionError::syscall_target_not_in_kernel(callee_hash, &err_ctx));
             }
 
             // set the system registers to the syscall context
@@ -520,7 +545,7 @@ impl FastProcessor {
         // when returning from a function call or a syscall, restore the context of the
         // system registers and the operand stack to what it was prior to
         // the call.
-        self.restore_context()?;
+        self.restore_context(&err_ctx)?;
 
         // Corresponds to the row inserted for the END operation added to the trace.
         self.clk += 1_u32;
@@ -543,11 +568,13 @@ impl FastProcessor {
             return Err(ExecutionError::CallInSyscall("dyncall"));
         }
 
+        let err_ctx = err_ctx!(program, dyn_node, self.source_manager.clone());
+
         // Retrieve callee hash from memory, using stack top as the memory address.
         let callee_hash = {
             let mem_addr = self.stack_get(0);
             self.memory
-                .read_word(self.ctx, mem_addr, self.clk)
+                .read_word(self.ctx, mem_addr, self.clk, &err_ctx)
                 .map_err(ExecutionError::MemoryError)?
         };
 
@@ -571,13 +598,13 @@ impl FastProcessor {
             None => {
                 let mast_forest = host
                     .get_mast_forest(&callee_hash)
-                    .ok_or_else(|| ExecutionError::dynamic_node_not_found(callee_hash, &()))?;
+                    .ok_or_else(|| ExecutionError::dynamic_node_not_found(callee_hash, &err_ctx))?;
 
                 // We limit the parts of the program that can be called externally to procedure
                 // roots, even though MAST doesn't have that restriction.
                 let root_id = mast_forest
                     .find_procedure_root(callee_hash)
-                    .ok_or(ExecutionError::malfored_mast_forest_in_host(callee_hash, &()))?;
+                    .ok_or(ExecutionError::malfored_mast_forest_in_host(callee_hash, &err_ctx))?;
 
                 self.execute_mast_node(root_id, &mast_forest, kernel, host)?
             },
@@ -585,7 +612,7 @@ impl FastProcessor {
 
         // For dyncall, restore the context.
         if dyn_node.is_dyncall() {
-            self.restore_context()?;
+            self.restore_context(&err_ctx)?;
         }
 
         // Corresponds to the row inserted for the END operation added to the trace.
@@ -608,6 +635,7 @@ impl FastProcessor {
     // Note: when executing individual ops, we do not increment the clock by 1 at every iteration
     // for performance reasons (~25% performance drop). Hence, `self.clk` cannot be used directly to
     // determine the number of operations executed in a program.
+    #[inline(always)]
     fn execute_basic_block_node(
         &mut self,
         basic_block_node: &BasicBlockNode,
@@ -624,6 +652,7 @@ impl FastProcessor {
         // execute first op batch
         if let Some(first_op_batch) = op_batches.next() {
             self.execute_op_batch(
+                basic_block_node,
                 first_op_batch,
                 &mut decorator_ids,
                 batch_offset_in_block,
@@ -639,6 +668,7 @@ impl FastProcessor {
             self.clk += 1_u32;
 
             self.execute_op_batch(
+                basic_block_node,
                 op_batch,
                 &mut decorator_ids,
                 batch_offset_in_block,
@@ -671,6 +701,7 @@ impl FastProcessor {
     #[inline(always)]
     fn execute_op_batch(
         &mut self,
+        basic_block: &BasicBlockNode,
         batch: &OpBatch,
         decorators: &mut DecoratorIterator,
         batch_offset_in_block: usize,
@@ -699,7 +730,10 @@ impl FastProcessor {
             }
 
             // decode and execute the operation
-            self.execute_op(op, batch_offset_in_block + op_idx_in_batch, program, host)?;
+            let op_idx_in_block = batch_offset_in_block + op_idx_in_batch;
+            let err_ctx =
+                err_ctx!(program, basic_block, self.source_manager.clone(), op_idx_in_block);
+            self.execute_op(op, op_idx_in_block, program, host, &err_ctx)?;
 
             // if the operation carries an immediate value, the value is stored at the next group
             // pointer; so, we advance the pointer to the following group
@@ -762,12 +796,14 @@ impl FastProcessor {
         Ok(())
     }
 
+    #[inline(always)]
     fn execute_op(
         &mut self,
         operation: &Operation,
         op_idx: usize,
         program: &MastForest,
         host: &mut impl Host,
+        err_ctx: &impl ErrorContext,
     ) -> Result<(), ExecutionError> {
         if self.bounds_check_counter == 0 {
             let err_str = if self.stack_top_idx - MIN_STACK_DEPTH == 0 {
@@ -783,13 +819,15 @@ impl FastProcessor {
             Operation::Noop => {
                 // do nothing
             },
-            Operation::Assert(err_code) => self.op_assert(*err_code, op_idx, host, program)?,
+            Operation::Assert(err_code) => {
+                self.op_assert(*err_code, op_idx, host, program, err_ctx)?
+            },
             Operation::FmpAdd => self.op_fmpadd(),
             Operation::FmpUpdate => self.op_fmpupdate()?,
             Operation::SDepth => self.op_sdepth(),
             Operation::Caller => self.op_caller()?,
             Operation::Clk => self.op_clk(op_idx)?,
-            Operation::Emit(event_id) => self.op_emit(*event_id, op_idx, host)?,
+            Operation::Emit(event_id) => self.op_emit(*event_id, op_idx, host, err_ctx)?,
 
             // ----- flow control operations ------------------------------------------------------
             // control flow operations are never executed directly
@@ -810,27 +848,27 @@ impl FastProcessor {
             Operation::Add => self.op_add()?,
             Operation::Neg => self.op_neg()?,
             Operation::Mul => self.op_mul()?,
-            Operation::Inv => self.op_inv(op_idx)?,
+            Operation::Inv => self.op_inv(op_idx, err_ctx)?,
             Operation::Incr => self.op_incr()?,
-            Operation::And => self.op_and()?,
-            Operation::Or => self.op_or()?,
-            Operation::Not => self.op_not()?,
+            Operation::And => self.op_and(err_ctx)?,
+            Operation::Or => self.op_or(err_ctx)?,
+            Operation::Not => self.op_not(err_ctx)?,
             Operation::Eq => self.op_eq()?,
             Operation::Eqz => self.op_eqz()?,
             Operation::Expacc => self.op_expacc(),
             Operation::Ext2Mul => self.op_ext2mul(),
 
             // ----- u32 operations ---------------------------------------------------------------
-            Operation::U32split => self.op_u32split()?,
-            Operation::U32add => self.op_u32add()?,
-            Operation::U32add3 => self.op_u32add3()?,
-            Operation::U32sub => self.op_u32sub(op_idx)?,
-            Operation::U32mul => self.op_u32mul()?,
-            Operation::U32madd => self.op_u32madd()?,
-            Operation::U32div => self.op_u32div(op_idx)?,
-            Operation::U32and => self.op_u32and()?,
-            Operation::U32xor => self.op_u32xor()?,
-            Operation::U32assert2(err_code) => self.op_u32assert2(*err_code)?,
+            Operation::U32split => self.op_u32split(),
+            Operation::U32add => self.op_u32add(err_ctx)?,
+            Operation::U32add3 => self.op_u32add3(err_ctx)?,
+            Operation::U32sub => self.op_u32sub(op_idx, err_ctx)?,
+            Operation::U32mul => self.op_u32mul(err_ctx)?,
+            Operation::U32madd => self.op_u32madd(err_ctx)?,
+            Operation::U32div => self.op_u32div(op_idx, err_ctx)?,
+            Operation::U32and => self.op_u32and(err_ctx)?,
+            Operation::U32xor => self.op_u32xor(err_ctx)?,
+            Operation::U32assert2(err_code) => self.op_u32assert2(*err_code, err_ctx)?,
 
             // ----- stack manipulation -----------------------------------------------------------
             Operation::Pad => self.op_pad(),
@@ -866,28 +904,28 @@ impl FastProcessor {
             Operation::MovDn6 => self.rotate_right(7),
             Operation::MovDn7 => self.rotate_right(8),
             Operation::MovDn8 => self.rotate_right(9),
-            Operation::CSwap => self.op_cswap()?,
-            Operation::CSwapW => self.op_cswapw()?,
+            Operation::CSwap => self.op_cswap(err_ctx)?,
+            Operation::CSwapW => self.op_cswapw(err_ctx)?,
 
             // ----- input / output ---------------------------------------------------------------
             Operation::Push(element) => self.op_push(*element),
-            Operation::AdvPop => self.op_advpop(op_idx, host)?,
-            Operation::AdvPopW => self.op_advpopw(op_idx, host)?,
-            Operation::MLoadW => self.op_mloadw(op_idx)?,
-            Operation::MStoreW => self.op_mstorew(op_idx)?,
-            Operation::MLoad => self.op_mload()?,
-            Operation::MStore => self.op_mstore()?,
-            Operation::MStream => self.op_mstream(op_idx)?,
-            Operation::Pipe => self.op_pipe(op_idx, host)?,
+            Operation::AdvPop => self.op_advpop(op_idx, host, err_ctx)?,
+            Operation::AdvPopW => self.op_advpopw(op_idx, host, err_ctx)?,
+            Operation::MLoadW => self.op_mloadw(op_idx, err_ctx)?,
+            Operation::MStoreW => self.op_mstorew(op_idx, err_ctx)?,
+            Operation::MLoad => self.op_mload(err_ctx)?,
+            Operation::MStore => self.op_mstore(err_ctx)?,
+            Operation::MStream => self.op_mstream(op_idx, err_ctx)?,
+            Operation::Pipe => self.op_pipe(op_idx, host, err_ctx)?,
 
             // ----- cryptographic operations -----------------------------------------------------
             Operation::HPerm => self.op_hperm(),
-            Operation::MpVerify(err_code) => self.op_mpverify(*err_code, host, program)?,
-            Operation::MrUpdate => self.op_mrupdate(host)?,
+            Operation::MpVerify(err_code) => self.op_mpverify(*err_code, host, program, err_ctx)?,
+            Operation::MrUpdate => self.op_mrupdate(host, err_ctx)?,
             Operation::FriE2F4 => self.op_fri_ext2fold4()?,
-            Operation::HornerBase => self.op_horner_eval_base(op_idx)?,
-            Operation::HornerExt => self.op_horner_eval_ext(op_idx)?,
-            Operation::ArithmeticCircuitEval => self.arithmetic_circuit_eval(op_idx)?,
+            Operation::HornerBase => self.op_horner_eval_base(op_idx, err_ctx)?,
+            Operation::HornerExt => self.op_horner_eval_ext(op_idx, err_ctx)?,
+            Operation::ArithmeticCircuitEval => self.arithmetic_circuit_eval(op_idx, err_ctx)?,
         }
 
         Ok(())
@@ -983,10 +1021,10 @@ impl FastProcessor {
     /// # Errors
     /// - Returns an error if the overflow stack is larger than the space available in the stack
     ///   buffer.
-    fn restore_context(&mut self) -> Result<(), ExecutionError> {
+    fn restore_context(&mut self, err_ctx: &impl ErrorContext) -> Result<(), ExecutionError> {
         // when a call/dyncall/syscall node ends, stack depth must be exactly 16.
         if self.stack_size() > MIN_STACK_DEPTH {
-            return Err(ExecutionError::invalid_stack_depth_on_return(self.stack_size(), &()));
+            return Err(ExecutionError::invalid_stack_depth_on_return(self.stack_size(), err_ctx));
         }
 
         let ctx_info = self